requirements.txt
## Ubuntu 16.04 recommended

django==1.11.3
djangorestframework==3.6.3
markdown==2.6.9
django-filter==1.1.0
django-rest-auth==0.9.2
fcm-django==0.2.12
jsonfield==2.0.2
django-allauth==0.34.0
simple_salesforce==0.73.0   # whys this here?
django-cors-headers==2.1.0
django-suit==0.2.25
django-polymorphic==1.3
django-rest-swagger==2.1.2
redis==2.10.5
Pillow==4.3.0
django-redis-sessions==0.5.6
jsbeautifier==1.7.4
drfdocs==0.0.11   # Another lib thats pretty much useless
coreapi==2.3.3
pygments==2.2.0
django-angular==1.1.4   # Dont really need this
gevent==1.2.2
psycogreen==1.0
django-grappelli==2.9.1
django-filebrowser==3.9.1
django-websocket-redis==0.5.1
django-redis==4.8.0
django_extensions==1.9.8
